requirements.txt
distlib==0.3.8
filelock==3.13.1
numpy==1.26.4
pandas==2.2.0
platformdirs==4.2.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2024.1
virtualenv==20.25.0
